site/backend/requirements.txt
python-dotenv
flask-sqlalchemy
flask-migrate
flask-wtf
flask-login
flask-rbac
Flask-Reuploaded
ffmpeg-python
opencv-python
pyee
Flask-Mail
raven
mysqlclient
gunicorn
gevent
email_validator
psycopg2-binary
